Bug#35081930 Fix compiler warnings on FreeBSD / clang 13
Fix some compiler warnings:

storage/ndb/src/common/portlib/NdbHW.cpp:1629:9:
error: unused variable 'cpu_sockets' [-Werror,-Wunused-variable]
  Int32 cpu_sockets;

storage/ndb/src/common/portlib/NdbHW.cpp:1654:24:
error: comparison of integers of different signs: 'Uint32' (aka 'unsigned int') and 'Int32' (aka 'int') [-Werror,-Wsign-compare]
  for (Uint32 i = 0; i < active_cpu; i++)

storage/ndb/src/common/transporter/SHM_Transporter.unix.cpp:49:5: error: format specifies type 'int' but the argument has type 'key_t' (aka 'long') [-Werror,-Wformat]

plugin/thread_pool/src/thread_pool.cc:863:41: error: format specifies type 'unsigned long' but the argument has type 'std::chrono::duration<long long, std::ratio<1, 1000>>::rep' (aka 'long long') [-Werror,-Wformat]

(cherry picked from commit 9f8f3db8defd31dd15c9f3579ad11b81846feba7)

+++ storage/ndb/src/common/transporter/SHM_Transporter.unix.cpp
@@ -46,7 +46,7 @@
 void SHM_Transporter::make_error_info(char info[], int sz)
 {
   snprintf(info,sz,"Shm key=%d sz=%d id=%d",
-	   shmKey, shmSize, shmId);
+           static_cast<int>(shmKey), shmSize, shmId);
 }
 
 bool
